Adds check for `patternMismatch`
